+++ app/src/main/java/com/ourstilt/customViews/animatedbottombar/AnimatedBottomBar.kt
@@ -1,4 +1,4 @@
-package com.ourstilt.animatedbottombar
+package com.ourstilt.customViews.animatedbottombar
 
 import android.content.Context
 import android.content.res.TypedArray
