+++ app/src/main/java/com/ourstilt/customViews/animatedbottombar/BottomBarStyle.kt
@@ -1,4 +1,4 @@
-package com.ourstilt.animatedbottombar
+package com.ourstilt.customViews.animatedbottombar
 
 import android.graphics.Color
 import android.graphics.Typeface
