+++ app/src/main/java/com/ourstilt/customViews/animatedbottombar/SavedState.kt
@@ -1,4 +1,4 @@
-package com.ourstilt.animatedbottombar
+package com.ourstilt.customViews.animatedbottombar
 
 import android.os.Parcel
 import android.os.Parcelable
