+++ app/src/main/java/com/ourstilt/homepage/HomeActivity.kt
@@ -23,7 +23,7 @@ import com.bumptech.glide.load.DataSource
 import com.bumptech.glide.load.engine.GlideException
 import com.bumptech.glide.request.RequestListener
 import com.ourstilt.R
-import com.ourstilt.animatedbottombar.AnimatedBottomBar
+import com.ourstilt.customViews.animatedbottombar.AnimatedBottomBar
 import com.ourstilt.base.BaseViewPagerAdapter
 import com.ourstilt.common.fadeIn
 import com.ourstilt.common.fadeOut
